requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
av==12.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
colorlog==6.8.2
comet-ml==3.44.4
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
cryptography==43.0.0
cycler==0.12.1
dash==2.17.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
docker-pycreds==0.4.0
docstring_parser==0.16
dulwich==0.22.1
embreex==2.17.7.post5
everett==3.1.0
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
fonttools==4.53.1
fqdn==1.5.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.35.0
importlib_metadata==8.2.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jaxtyping==0.2.33
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lightning-utilities==0.11.6
lit==18.1.8
lxml==5.3.0
manifold3d==2.5.1
mapbox_earcut==1.0.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy==0.4.8
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nerfacc==0.5.2
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nodeenv==1.9.1
notebook_shim==0.2.4
numpy==1.26.4
nuscenes-devkit==1.1.11
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.6.0.66
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.2.2
plotly==5.23.0
plyfile==0.8.1
prometheus_client==0.20.0
promise==2.3
prompt_toolkit==3.0.47
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pybind11==2.13.3
pycocotools==2.0.8
pycollada==0.8
pycolmap==3.10.0
pycparser==2.22
Pygments==2.18.0
pyliblzfse==0.4.1
pymeshlab==2023.12.post1
pyngrok==7.2.0
pyparsing==3.1.2
pyquaternion==0.9.9
PySocks==1.7.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-json-logger==2.0.7
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
rawpy==0.18.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
Rtree==1.3.0
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.13.0
setproctitle==1.3.3
Shapely==1.8.5.post1
shtab==1.7.1
simple-websocket==1.0.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
splines==0.3.0
stack-data==0.6.3
svg.path==6.3
sympy==1.13.2
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.8.10
timm==0.6.7
tinycss2==1.3.0
tomli==2.0.1
torch==2.0.1+cu118
torch-fidelity==0.3.0
torchmetrics==1.4.1
torchvision==0.15.2+cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trimesh==4.4.4
triton==2.0.0
typed-argument-parser==1.10.1
typeguard==2.13.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.6
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
vhacdx==0.0.8.post1
viser==0.1.3
wandb==0.17.6
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
wsproto==1.2.0
wurlitzer==3.1.1
xatlas==0.0.9
xxhash==3.4.1
yourdfpy==0.0.56
zipp==3.20.0
